Bug#553616: [cowbuilder] cannot lock DBdriver config during parallel usage
Package: cowbuilder
Version: 0.57
Severity: normal

I noticed a quite bad behavior when building multiple packages at the same
time. The problem seems to be /var/cache/debconf/config.dat. Multiple packages
want to install x11-common which uses these file over debconf. This means that
this file gets locked (i would guess using flock). As this file is a hardlink
all other files (which are only hardlinks) will be locked too and the install
fails. This means that the whole build fails due to the usage of cowbuilder.
This means that cowdancer must also copy files on flock and not only on
writes.

I think that it has the same problems as fchmod because we are doing it at
file table level. Maybe a workaround for known debian specific locked files
can be added before adding cowdancer overrides.

Just added a warning stub.