Bug#597102: cowdancer: diff for NMU version 0.62+nmu2
tags 597102 + pending
thanks

Dear maintainer,

I've prepared an NMU for cowdancer (versioned as 0.62+nmu2) and
uploaded it to DELAYED/2. Please feel free to tell me if I
should delay it longer.

Sorry, when preparing the NMU for #564425 I didn't noticed, that it can't
migrate as this would introduce #597102.  So here I am again :(

Regards.