Bug#648438: cowdancer: FTBFS with ld that defaults to --as-needed: bad link order
Package: cowdancer
Version: 0.65
Severity: important
Tags: patch
User: ubuntu-devel@lists.ubuntu.com
Usertags: origin-ubuntu ubuntu-patch oneiric

cowdancer fails to build with a linker that defaults to --as-needed, as
shown in this Ubuntu build log:

  https://launchpadlibrarian.net/83906438/buildlog_ubuntu-precise-i386.cowdancer_0.65_FAILEDTOBUILD.txt.gz

This is because of incorrect link ordering: it puts libraries before the
objects that use them rather than after.  See:

  http://wiki.debian.org/ToolChain/DSOLinking#Only_link_with_needed_libraries

The following patch fixes this.

  * Fix link order to list libraries after the objects that require them.
  * Use -pthread rather than -lpthread, since it should be a superset, it's
    what's recommended in pthread_create(3), and it isn't so sensitive to
    link order.